+++ th2/data_services/exceptions.py
@@ -12,27 +12,30 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 class EventNotFound(Exception):
-    def __init__(self, id_):
+    def __init__(self, id_, error_description):
         """Exception for the case when the the event was not found in data source.
         Args:
             id_: Event id.
+            error_description: Description of error
         """
-        self._id = id_
+        self._error_description = error_description
 
     def __str__(self):
-        return f"Unable to find the event with id '{self._id}'"
+        return "An error occured while trying to get event with id: {self._id}. Description of error: {self._error_description}"
 
 
 class MessageNotFound(Exception):
-    def __init__(self, id_):
+    def __init__(self, id_, error_description):
         """Exception for the case when the the message was not found in data source.
         Args:
             id_: Event id.
+            error_description: Description of error
         """
         self._id = id_
+        self._error_description = error_description
 
     def __str__(self):
-        return f"Unable to find the message with id '{self._id}'"
+        return "An error occured while trying to get message with id: {self._id}. Description of error: {self._error_description}"
 
 
 class CommandError(Exception):